backend/requirements.txt
# Django Backend Dependencies
# Core Django and REST Framework
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.63.3
cryptography==41.0.3

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# API Documentation
drf-spectacular==0.27.2

# Development & Debugging
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Testing
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0

# Utilities
Pillow==10.0.0
python-slugify==8.0.1
redis==4.6.0

# Linting & Formatting
black==23.7.0
flake8==6.0.0
isort==5.12.0
